PC98 RBYTE image encoder
Added `rbyte_enc.py`, an encoder (stand-alone for now) for the PC98 flavor of the "RBYTE" image format. Supply an image (e.g. FILE.PNG) filename on the command line. This will generate FILE_rbyte.bin.

If you give an image
- ... using the 8-color RGB111 palette
- ... with dimensions 640x200 or less
- ... where the width is a multiple of 8 pixels
- ... vertically doubled (so, 640x400 or less)
... then this encoder should give you RBYTE data for it that is as small as possible and decodes to an identical-looking image.

Currently the optimizer makes the same decisions apparent in Lafiel Soft's PC98 RBYTE data - sometimes these result in slower-than-expected decoding since fancy compressed lines are preferred to uncompressed ones when the resulting number of bytes in the file would be the same.

On the other hand...
- If your image is larger than 640x400 it will be scaled down to fit in that size, preserving the aspect ratio.
- If your image is not scan-doubled (that is, pairs of consecutive scanlines are not identical), an arbitrary pixel will be chosen from each vertical pair of pixels.
- If your image width isn't a multiple of 8 pixels the encoder will add enough black pixels on the right to reach an 8px boundary.
- If the image contains colors outside the RGB111 palette, they will be forced into RGB111 by crude stippling.